+++ misc/update_DESCRIPTION_NEWS.R
@@ -301,7 +301,9 @@ my_news$add_bullet(c("including GitHub Actions",
                      "adjusting tests for Windows",
                      "adjusting needed R version"))
 
-# set R version
+
+# set R version -----------------------------------------------------------
+
 my_desc$set_dep("R", type = desc::dep_types[2], version = ">= 3.5")
 
 my_news$add_subtitle("styling")
@@ -316,6 +318,15 @@ my_news$add_subtitle("removing dependencies")
 my_news$add_bullet("RCurl, stringr")
 my_desc$del_dep("RCurl", type = desc::dep_types[1])
 my_desc$del_dep("stringr", type = desc::dep_types[1])
+
+
+# adjusting test setup ----------------------------------------------------
+
+my_desc$bump_version("patch")
+my_news$add_version(my_desc$get_version())
+my_news$add_subtitle("internal test setup changes")
+my_news$add_bullet(c("removing ubuntu with R version 4.0 in the test setup",
+                     "changing lintr options"))
 
 # save everything ---------------------------------------------------------